Refactor callbacks to unified core controller
