+++ analytics/anomaly_detection/analyzer.py
@@ -12,10 +12,7 @@ from sklearn.preprocessing import StandardScaler
 from sklearn.exceptions import DataConversionWarning
 import logging
 import warnings
-from services.data_processing.callback_controller import (
-    SecurityEvent,
-    emit_security_event,
-)
+from core.callback_controller import CallbackEvent as SecurityEvent, fire_event as emit_security_event
 
 from .types import AnomalyAnalysis
 from .data_prep import prepare_anomaly_data
