+++ core/unicode_utils.py
@@ -1,0 +1,4 @@
+"""Compatibility wrapper importing unicode utilities."""
+
+from utils.unicode_utils import *  # noqa: F401,F403
+
