+++ pages/deep_analytics/__init__.py
@@ -29,7 +29,7 @@ from services.data_processing.analytics_engine import (
     process_suggests_analysis_safe,
 )
 from utils.unicode_utils import sanitize_dataframe
-from services.data_processing.callback_controller import CallbackController
+from core.callback_controller import CallbackController
 
 __all__ = [
     "layout",
